refactor(watch): move to runtime-html
